backend/requirements.txt
django==2.2.28
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
Pillow==9.2.0